Updated Player Name and Jersey Number

+++ PIHLSite/Controllers/PenaltyRecordController.cs
@@ -51,7 +51,7 @@ namespace PIHLSite.Controllers
         {
             ViewData["GameId"] = new SelectList(_context.Games, "GameId", "GameId");
             ViewData["PenaltyId"] = new SelectList(_context.Penalties, "PenaltyId", "PenaltyDescription");
-            ViewData["PlayerId"] = new SelectList(_context.Players, "PlayerId", "LastName");
+            ViewData["PlayerId"] = new SelectList(_context.Players, "PlayerId", "NameandNumber");
             return View();
         }
 
@@ -67,6 +67,15 @@ namespace PIHLSite.Controllers
             if (gameRecord == null)
             {
                 return NotFound();
+            }
+            TimeSpan duration = new TimeSpan(0,60,0);
+            if (penaltyRecord.Pim > duration)
+            {
+                TempData["Message"] = "The time of the penalty cannot be greater than a game misconduct (60 minutes)";
+                ViewData["GameId"] = new SelectList(_context.Games, "GameId", "GameId");
+                ViewData["PenaltyId"] = new SelectList(_context.Penalties, "PenaltyId", "PenaltyDescription");
+                ViewData["PlayerId"] = new SelectList(_context.Players, "PlayerId", "NameandNumber");
+                return View();
             }
             if (playerRecord != null)
             {
@@ -84,13 +93,13 @@ namespace PIHLSite.Controllers
                 }
                 catch (Exception ex)
                 {
-                    return RedirectToAction(nameof(Index));
+                    return Redirect(Url.Action("Index", "Scorekeeper"));
                 }
-                return RedirectToAction(nameof(Index));
+                return Redirect(Url.Action("Index", "Scorekeeper"));
             }
             ViewData["GameId"] = new SelectList(_context.Games, "GameId", "GameId", penaltyRecord.GameId);
             ViewData["PenaltyId"] = new SelectList(_context.Penalties, "PenaltyId", "PenaltyDescription", penaltyRecord.PenaltyId);
-            ViewData["PlayerId"] = new SelectList(_context.Players, "PlayerId", "LastName", penaltyRecord.PlayerId);
+            ViewData["PlayerId"] = new SelectList(_context.Players, "PlayerId", "NameandNumber", penaltyRecord.PlayerId);
             return View(penaltyRecord);
         }
 
@@ -110,7 +119,7 @@ namespace PIHLSite.Controllers
             }
             ViewData["GameId"] = new SelectList(_context.Games, "GameId", "GameId", penaltyRecord.GameId);
             ViewData["PenaltyId"] = new SelectList(_context.Penalties, "PenaltyId", "PenaltyDescription", penaltyRecord.PenaltyId);
-            ViewData["PlayerId"] = new SelectList(_context.Players, "PlayerId", "LastName", penaltyRecord.PlayerId);
+            ViewData["PlayerId"] = new SelectList(_context.Players, "PlayerId", "NameandNumber", penaltyRecord.PlayerId);
             return View(penaltyRecord);
         }
 
@@ -148,13 +157,14 @@ namespace PIHLSite.Controllers
             }
             ViewData["GameId"] = new SelectList(_context.Games, "GameId", "GameId", penaltyRecord.GameId);
             ViewData["PenaltyId"] = new SelectList(_context.Penalties, "PenaltyId", "PenaltyCode", penaltyRecord.PenaltyId);
-            ViewData["PlayerId"] = new SelectList(_context.Players, "PlayerId", "FirstName", penaltyRecord.PlayerId);
+            ViewData["PlayerId"] = new SelectList(_context.Players, "PlayerId", "NameandNumber", penaltyRecord.PlayerId);
             return View(penaltyRecord);
         }
 
         // GET: PenaltyRecord/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+
             if (id == null)
             {
                 return NotFound();
@@ -179,9 +189,29 @@ namespace PIHLSite.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var penaltyRecord = await _context.PenaltyRecords.FindAsync(id);
-            _context.PenaltyRecords.Remove(penaltyRecord);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var gameRecord = await _context.Set<Game>().FirstOrDefaultAsync(o => o.GameId == penaltyRecord.GameId);
+            var playerRecord = await _context.Set<Player>().FirstOrDefaultAsync(o => o.PlayerId == penaltyRecord.PlayerId);
+            if (gameRecord == null)
+            {
+                return NotFound();
+            }
+            if (playerRecord != null)
+            {
+                playerRecord.Pimtotal -= penaltyRecord.Pim;
+            }
+            try
+            {
+                _context.PenaltyRecords.Remove(penaltyRecord);
+                await _context.SaveChangesAsync();
+
+                _context.Players.Add(playerRecord);
+                await _context.SaveChangesAsync();
+            }
+            catch (Exception)
+            {
+                return Redirect(Url.Action("Index", "Scorekeeper"));
+            }
+            return Redirect(Url.Action("Index", "Scorekeeper"));
         }
 
         private bool PenaltyRecordExists(int id)
